Compute real and coreset costs.

+++ xrun/calc_costs.py
@@ -119,7 +119,7 @@ def load_covertype_dataset():
 
 datasets = dict()
 
-def load_raw_data(coreset_file_path: Path):
+def load_original_data(coreset_file_path: Path):
     dataset, algorithm, k = re.findall(r"/.+/(.+)/(.+)-k(\d+)-", str(coreset_file_path))[0]
     loaders = {
         "census": load_census_data,
@@ -136,13 +136,13 @@ def load_raw_data(coreset_file_path: Path):
     return datasets[dataset]
 
 
-def compute_weighted_cost(data_file_path: Path, centers_file_path: Path) -> Path:
-    cost_file_path = data_file_path.parent / "weighted_cost.txt"
+def compute_real_cost(coreset_file_path: Path, centers_file_path: Path) -> Path:
+    cost_file_path = coreset_file_path.parent / "real_cost.txt"
     if cost_file_path.exists():
         return cost_file_path
 
-    print("Computing weighted cost...")
-    data_points = load_raw_data(data_file_path)
+    print("Computing real cost... ", end="")
+    data_points = load_original_data(coreset_file_path)
 
     centers = np.loadtxt(fname=centers_file_path, dtype=np.double, delimiter=' ', skiprows=0)
     center_weights = centers[:,0] 
@@ -155,9 +155,39 @@ def compute_weighted_cost(data_file_path: Path, centers_file_path: Path) -> Path
     dist_closest_centers = np.min(D, axis=1)
 
     # Weigh the distances and sum it all up
-    weighted_cost = np.sum(dist_closest_centers)
+    cost = np.sum(dist_closest_centers)
 
-    print(f"Computed weighted cost: {weighted_cost}")
+    print(f"Computed real cost: {cost}")
+    
+    with open(cost_file_path, "w") as f:
+        f.write(str(cost))
+    return cost_file_path
+
+
+def compute_coreset_costs(coreset_file_path: Path, centers_file_path: Path) -> Path:
+    cost_file_path = coreset_file_path.parent / "coreset_cost.txt"
+    if cost_file_path.exists():
+        return cost_file_path
+
+    print("Computing coreset cost... ", end='')
+    coreset = np.loadtxt(fname=coreset_file_path, dtype=np.double, delimiter=' ', skiprows=1)
+    coreset_weights = coreset[:,0]
+    coreset_points = coreset[:,1:]
+
+    centers = np.loadtxt(fname=centers_file_path, dtype=np.double, delimiter=' ', skiprows=0)
+    center_weights = centers[:,0] 
+    center_points = centers[:,1:]
+
+    # Distances between all corset points and center points
+    D = pairwise_distances(coreset_points, center_points, metric="sqeuclidean")
+
+    # For each point (w, p) in S, find the distance to its closest center
+    dist_closest_centers = np.min(D, axis=1)
+
+    # Weigh the distances and sum it all up
+    weighted_cost = np.sum(coreset_weights * dist_closest_centers)
+
+    print(f"Computed coreset cost: {weighted_cost}")
     
     with open(cost_file_path, "w") as f:
         f.write(str(weighted_cost))
@@ -180,7 +210,8 @@ def main(results_dir: str) -> None:
         print(f"Processing file {index+1} of {total_files}...")
         data_file_path = unzip_file(file_path)
         centers_file_path = compute_centers(data_file_path)
-        cost_file_path = compute_weighted_cost(data_file_path, centers_file_path)
+        compute_real_cost(data_file_path, centers_file_path)
+        compute_coreset_costs(data_file_path, centers_file_path)
 
 
 if __name__ == "__main__":
